More Refactoring

+++ src/main/java/br/com/ab/Trello/ws/CommentSoapWS.java
@@ -1,36 +1,51 @@
-package br.com.ab.Trello.service;
+package br.com.ab.Trello.ws;
 
 import java.util.List;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import javax.jws.WebParam;
+import javax.jws.WebResult;
+import javax.jws.WebService;
+import javax.xml.bind.annotation.XmlElement;
 
 import br.com.ab.Trello.dao.CommentDao;
 import br.com.ab.Trello.model.Comment;
 
+@WebService
 @Stateless
-public class CommentService {
+public class CommentSoapWS {
 
 	@Inject
 	CommentDao commentDao;
 	
-	public void addComment(Comment comment){
-		this.commentDao.addComment(comment);
+	public void addComment(Comment comment) {
+		if (!comment.equals(null)) {
+			this.commentDao.addComment(comment);
+		}
 	}
-	
-	public Comment findCommentById(Integer commentId){
+
+	@WebResult(name = "commentFound")
+	public Comment findCommentById(
+			@WebParam(name = "commentId") @XmlElement(required = true, nillable = false) Integer commentId) {
 		return this.commentDao.findCommentById(commentId);
 	}
-	
-	public List<Comment> findAllComments(){
+
+	@WebResult(name = "allComments")
+	public List<Comment> findAllComments() {
 		return this.commentDao.findAllComments();
 	}
-	
-	public List<Comment> findAllCommentsFromCard(Integer cardId){
+
+	@WebResult(name = "allComments")
+	public List<Comment> findAllCommentsFromCard(
+			@WebParam(name = "cardId") @XmlElement(required = true, nillable = false) Integer cardId) {
 		return this.commentDao.findAllCommentsFromCardId(cardId);
 	}
-	
-	public void deleteComment(Comment comment){
-		this.commentDao.deleteComment(comment);
+
+	public void deleteComment(
+			@WebParam(name = "comment") @XmlElement(required = true, nillable = false) Comment comment) {
+		if (!comment.equals(null)) {
+			this.commentDao.deleteComment(comment);
+		}
 	}
 }
